sony-blr-agentic-ai-practices/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental

python-dotenv
ipykernel
pydantic
numexpr

faiss-cpu
unstructured
langchain-pinecone
pinecone-client
streamlit

langchain-ollama
langdetect
deep-translator

gradio

tabulate

wikipedia
arxiv
langgraph
langchain-anthropic

chroma
langchain-chroma
nbformat